requirements.txt

# Django
Django==1.9.5

# Plugins
django-filter==0.13.0         # [r:DRF] Filters for Django REST framework
markdown==2.6.6               # [r:DRF] Markdown. For HTML rendering of API views
djangorestframework==3.3.3          # Django REST framework
